requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
openai>=1.10.0
python-dotenv>=1.0.0
duckduckgo-search>=4.1.0
pypdf>=3.17.0
unstructured>=0.11.0
tiktoken>=0.5.0
chromadb>=0.4.22